Add css styles to all view,edit,delete buttons

+++ src/features/Course/CourseDetail.js
@@ -1,5 +1,4 @@
 import generalStyles from "../../generalStyles.module.css";
-import EditCourseForm from "../../components/Form/EditCourseForm";
 import styles from "../Profile.module.css";
 import { useParams, useNavigate } from "react-router-dom";
 import React, { useEffect, useState } from "react";
@@ -18,7 +17,7 @@ import formStyles from "../../components/Form/Form.module.css";
 function CourseDetail() {
   const { courseNo } = useParams();
   const navigate = useNavigate();
-  const [originalCourse, setOriginalCourse] = useState(null); 
+  const [originalCourse, setOriginalCourse] = useState(null);
   const [enrollments, setEnrollments] = useState([]);
 
   const [course, setCourse] = useState(null);
@@ -69,6 +68,11 @@ function CourseDetail() {
   }, [courseNo]);
 
   const handleDeleteCourse = async () => {
+    const confirmed = window.confirm(
+      "Are you sure you want to delete this course?"
+    );
+    if (!confirmed) return;
+
     try {
       await deleteCourse(course.CourseID);
       alert("Course deleted successfully!");
@@ -135,234 +139,249 @@ function CourseDetail() {
 
   return (
     <div>
-      <MainTitle title="Course Detail" />
-      {course ? (
-        <EditContainer
-          title={course.CourseName}
-          isEdit={isEditing}
-          onClickEdit={handleEditBtn}
-          onClickSave={handleClickSave}
-          onClickCancel={handleCancelEdit}
-        >
-          <div className={formStyles.sectionLayout}>
-            <form>
-              <div className={formStyles.formRow}>
-                <div className={formStyles.formItem}>
-                  <label htmlFor="courseID" className={formStyles.formLabel}>
-                    Course ID
-                  </label>
-                  <input
-                    type="text"
-                    id="courseID"
-                    name="CourseID"
-                    className={formStyles.formInput}
-                    readOnly
-                    disabled
-                    value={course.CourseID}
-                  />
-                </div>
-              </div>
-
-              <div className={formStyles.formRow}>
-                <div className={formStyles.formItem}>
-                  <label htmlFor="courseName" className={formStyles.formLabel}>
-                    Course Name
-                  </label>
-                  <input
-                    type="text"
-                    id="courseName"
-                    name="CourseName"
-                    className={formStyles.formInput}
-                    disabled={!isEditing}
-                    value={course.CourseName}
-                    onChange={handleChange}
-                  />
-                </div>
-                <div className={formStyles.formItem}>
-                  <label htmlFor="description" className={formStyles.formLabel}>
-                    Course Description
-                  </label>
-                  <input
-                    type="text"
-                    id="description"
-                    name="Description"
-                    className={formStyles.formInput}
-                    disabled={!isEditing}
-                    value={course.Description}
-                    onChange={handleChange}
-                  />
-                </div>
-              </div>
-
-              <div className={formStyles.formRow}>
-                <div className={formStyles.formItem}>
-                  <label htmlFor="startDate" className={formStyles.formLabel}>
-                    Start Date
-                  </label>
-                  <input
-                    type="date"
-                    id="startDate"
-                    name="StartDate"
-                    className={formStyles.formInput}
-                    disabled={!isEditing}
-                    value={course.StartDate}
-                    onChange={handleChange}
-                  />
-                </div>
-                <div className={formStyles.formItem}>
-                  <label htmlFor="endDate" className={formStyles.formLabel}>
-                    End Date
-                  </label>
-                  <input
-                    type="date"
-                    id="endDate"
-                    name="EndDate"
-                    className={formStyles.formInput}
-                    disabled={!isEditing}
-                    value={course.EndDate}
-                    onChange={handleChange}
-                  />
-                </div>
-
-                <div className={formStyles.formItem}>
-                  <label htmlFor="time" className={formStyles.formLabel}>
-                    Time
-                  </label>
-                  <input
-                    type="text"
-                    id="time"
-                    name="Time"
-                    className={formStyles.formInput}
-                    disabled={!isEditing}
-                    value={course.Time}
-                    onChange={handleChange}
-                  />
-                </div>
-              </div>
-
-              <div className={formStyles.formRow}>
-                <div className={formStyles.formItem}>
-                  <label htmlFor="teacherName" className={formStyles.formLabel}>
-                    Teacher Name
-                  </label>
-                  {isEditing ? (
-                    teachers && teachers.length > 0 ? (
-                      <select
-                        value={course.TeacherID}
-                        name="TeacherID"
-                        onChange={handleChange}
-                        required
-                        className={formStyles.formInput}
-                      >
-                        {teachers.map((teacher) => (
-                          <option
-                            key={teacher.TeacherID}
-                            value={teacher.TeacherID}
-                          >
-                            {teacher.Users.FirstName} {teacher.Users.LastName}
-                          </option>
-                        ))}
-                      </select>
-                    ) : (
-                      <p>No teachers found or loading...</p>
-                    )
-                  ) : (
+      <MainTitle
+        title={
+          isLoading ? `Course Detail` : `Course Detail: ${course?.CourseName}`
+        }
+        goBack={true}
+      />
+      <div className={styles.mainColumn}>
+        {course ? (
+          <EditContainer
+            title={course?.CourseName}
+            editButtonText="Edit"
+            isEdit={isEditing}
+            onClickEdit={handleEditBtn}
+            onClickSave={handleClickSave}
+            onClickCancel={handleCancelEdit}
+            onClickDelete={handleDeleteCourse}
+          >
+            <div className={formStyles.sectionLayout}>
+              <form>
+                <div className={formStyles.formRow}>
+                  <div className={formStyles.formItem}>
+                    <label htmlFor="courseID" className={formStyles.formLabel}>
+                      Course ID
+                    </label>
                     <input
                       type="text"
-                      id="teacherName"
-                      name="TeacherName"
+                      id="courseID"
+                      name="CourseID"
+                      className={formStyles.formInput}
+                      readOnly
+                      disabled
+                      value={course.CourseID}
+                    />
+                  </div>
+                </div>
+
+                <div className={formStyles.formRow}>
+                  <div className={formStyles.formItem}>
+                    <label
+                      htmlFor="courseName"
+                      className={formStyles.formLabel}
+                    >
+                      Course Name
+                    </label>
+                    <input
+                      type="text"
+                      id="courseName"
+                      name="CourseName"
                       className={formStyles.formInput}
                       disabled={!isEditing}
-                      value={`${course.TeacherUser.FirstName} ${course.TeacherUser.LastName}`}
+                      value={course.CourseName}
+                      onChange={handleChange}
                     />
-                  )}
+                  </div>
+                  <div className={formStyles.formItem}>
+                    <label
+                      htmlFor="description"
+                      className={formStyles.formLabel}
+                    >
+                      Course Description
+                    </label>
+                    <input
+                      type="text"
+                      id="description"
+                      name="Description"
+                      className={formStyles.formInput}
+                      disabled={!isEditing}
+                      value={course.Description}
+                      onChange={handleChange}
+                    />
+                  </div>
                 </div>
-              </div>
 
-              <div className={formStyles.formRow}>
-                <div className={formStyles.formItem}>
-                  <label htmlFor="programName" className={formStyles.formLabel}>
-                    Course Category
-                  </label>
-                  <input
-                    type="text"
-                    id="programName"
-                    name="ProgramName"
-                    className={formStyles.formInput}
-                    disabled
-                    value={course.Programs.ProgramName}
-                  />
-                </div>
-                <div className={formStyles.formItem}>
-                  <label htmlFor="programCode" className={formStyles.formLabel}>
-                    Course Category Code
-                  </label>
-                  <input
-                    type="text"
-                    id="programCode"
-                    name="ProgramCode"
-                    className={formStyles.formInput}
-                    disabled
-                    value={course.Programs.ProgramCode}
-                  />
-                </div>
-              </div>
-            </form>
-          </div>
+                <div className={formStyles.formRow}>
+                  <div className={formStyles.formItem}>
+                    <label htmlFor="startDate" className={formStyles.formLabel}>
+                      Start Date
+                    </label>
+                    <input
+                      type="date"
+                      id="startDate"
+                      name="StartDate"
+                      className={formStyles.formInput}
+                      disabled={!isEditing}
+                      value={course.StartDate}
+                      onChange={handleChange}
+                    />
+                  </div>
+                  <div className={formStyles.formItem}>
+                    <label htmlFor="endDate" className={formStyles.formLabel}>
+                      End Date
+                    </label>
+                    <input
+                      type="date"
+                      id="endDate"
+                      name="EndDate"
+                      className={formStyles.formInput}
+                      disabled={!isEditing}
+                      value={course.EndDate}
+                      onChange={handleChange}
+                    />
+                  </div>
 
-          <div className={formStyles.buttons}>
-            {!isEditing ? (
-              <Button onClickBtn={handleEditBtn}>Edit Course</Button>
-            ) : (
-              <>
-                <Button onClickBtn={handleClickSave}>Save</Button>
-                <Button onClickBtn={handleCancelEdit}>Cancel</Button>
-              </>
-            )}
-            <Button
-              onClickBtn={handleDeleteCourse}
-              className={styles.deleteBtn}
-            >
-              Delete Course
-            </Button>
-            <Button
-              onClickBtn={handleBack}
-              className={generalStyles.secondaryBtn}
-            >
-              Back to List
-            </Button>
+                  <div className={formStyles.formItem}>
+                    <label htmlFor="time" className={formStyles.formLabel}>
+                      Time
+                    </label>
+                    <input
+                      type="text"
+                      id="time"
+                      name="Time"
+                      className={formStyles.formInput}
+                      disabled={!isEditing}
+                      value={course.Time}
+                      onChange={handleChange}
+                    />
+                  </div>
+                </div>
+
+                <div className={formStyles.formRow}>
+                  <div className={formStyles.formItem}>
+                    <label
+                      htmlFor="teacherName"
+                      className={formStyles.formLabel}
+                    >
+                      Teacher Name
+                    </label>
+                    {isEditing ? (
+                      teachers && teachers.length > 0 ? (
+                        <select
+                          value={course.TeacherID}
+                          name="TeacherID"
+                          onChange={handleChange}
+                          required
+                          className={formStyles.formInput}
+                        >
+                          {teachers.map((teacher) => (
+                            <option
+                              key={teacher.TeacherID}
+                              value={teacher.TeacherID}
+                            >
+                              {teacher.Users.FirstName} {teacher.Users.LastName}
+                            </option>
+                          ))}
+                        </select>
+                      ) : (
+                        <p>No teachers found or loading...</p>
+                      )
+                    ) : (
+                      <input
+                        type="text"
+                        id="teacherName"
+                        name="TeacherName"
+                        className={formStyles.formInput}
+                        disabled={!isEditing}
+                        value={`${course.TeacherUser.FirstName} ${course.TeacherUser.LastName}`}
+                      />
+                    )}
+                  </div>
+                </div>
+
+                <div className={formStyles.formRow}>
+                  <div className={formStyles.formItem}>
+                    <label
+                      htmlFor="programName"
+                      className={formStyles.formLabel}
+                    >
+                      Course Category
+                    </label>
+                    <input
+                      type="text"
+                      id="programName"
+                      name="ProgramName"
+                      className={formStyles.formInput}
+                      disabled
+                      value={course.Programs.ProgramName}
+                    />
+                  </div>
+                  <div className={formStyles.formItem}>
+                    <label
+                      htmlFor="programCode"
+                      className={formStyles.formLabel}
+                    >
+                      Course Category Code
+                    </label>
+                    <input
+                      type="text"
+                      id="programCode"
+                      name="ProgramCode"
+                      className={formStyles.formInput}
+                      disabled
+                      value={course.Programs.ProgramCode}
+                    />
+                  </div>
+                </div>
+              </form>
+            </div>
+
+            {/* <div className={formStyles.buttons}>
+              {!isEditing ? (
+                <Button onClickBtn={handleEditBtn}>Edit Course</Button>
+              ) : (
+                <>
+                  <Button onClickBtn={handleClickSave}>Save</Button>
+                  <Button onClickBtn={handleCancelEdit}>Cancel</Button>
+                </>
+              )}
+              <Button onClickBtn={handleDeleteCourse} color="blue">
+                Delete Course
+              </Button>
+            </div> */}
+          </EditContainer>
+        ) : (
+          <div>No course data found.</div>
+        )}
+
+        <EditContainer title="Enrolled Students">
+          <div className={formStyles.formContainer}>
+            <table className={formStyles.courseTable}>
+              <thead>
+                <tr>
+                  <th>Student Name</th>
+                  <th>Email</th>
+                </tr>
+              </thead>
+              <tbody>
+                {enrollments.length > 0 ? (
+                  enrollments.map((enrollment) => (
+                    <tr key={enrollment.StudentID}>
+                      <td>{enrollment.StudentName}</td>
+                      <td>{enrollment.Email}</td>
+                    </tr>
+                  ))
+                ) : (
+                  <tr>
+                    <td colSpan="2">No students enrolled in this course.</td>
+                  </tr>
+                )}
+              </tbody>
+            </table>
           </div>
         </EditContainer>
-      ) : (
-        <div>No course data found.</div>
-      )}
-
-      <EditContainer title="Enrolled Students">
-        <div className={formStyles.formContainer}>
-          <table className={formStyles.courseTable}>
-            <thead>
-              <tr>
-                <th>Student Name</th>
-                <th>Email</th>
-              </tr>
-            </thead>
-            <tbody>
-              {enrollments.length > 0 ? (
-                enrollments.map((enrollment) => (
-                  <tr key={enrollment.StudentID}>
-                    <td>{enrollment.StudentName}</td>
-                    <td>{enrollment.Email}</td>
-                  </tr>
-                ))
-              ) : (
-                <tr>
-                  <td colSpan="2">No students enrolled in this course.</td>
-                </tr>
-              )}
-            </tbody>
-          </table>
-        </div>
-      </EditContainer>
+      </div>
     </div>
   );
 }
